refactor(ui): move diff rendering into diffrender.go
diffview.go keeps commands, message handlers, keys, cursor and
annotation state; every paint-path func now lives beside the
highlight cache it reads. Pure moves.
